feat: ajout de /home à l'url de la homepage connectée
